abstracted out bing base map options
